pyproject.toml
[project]
name = "bikeshed"
version = "0.1.0"
description = "MCP Playground"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "click>=8.1.8",
    "fastapi>=0.115.11",
    "fastapi-events>=0.12.2",
    "fasthx[jinja]>=2.2.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "litellm>=1.63.11",
    "loguru>=0.7.3",
    "markdown2>=2.5.3",
    "mcp>=1.3.0",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.8.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "rich>=13.9.4",
    "sse-starlette>=2.2.1",
    "transitions>=0.9.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "faker>=37.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-watch>=4.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
